[BE] Upload user avatar https://trello.com/c/pBNMU33u/76-be-upload-user-avatar
1.Added avatar size and thumbnail size in services.yaml (sent in function as constructor parameters)
2.Verified if the avatar is uploaded or not and handling this situation properly.
3.Images are now stored in public directory
4.Created UserAvatarManager class who takes care of writing and removing images from directories
5.Generalized crop and resize functions in order to be used in another further situations
6.Updated .gitignore file

+++ src/Transformer/UserTransformer.php
@@ -119,13 +119,15 @@ class UserTransformer
             }
         }
 
-        $userAvatar = new UploadedBase64EncodedFile(new Base64EncodedFile($dto->avatar));
+        if (!empty($dto->avatar)) {
+            $userAvatar = new UploadedBase64EncodedFile(new Base64EncodedFile($dto->avatar));
 
-        if ($user->getAvatar()) {
-            $this->imageRepository->removeUserAvatar($user);
+            if ($user->getAvatar()) {
+                $this->imageRepository->removeUserAvatar($user);
+            }
+
+            $this->imageRepository->uploadUserAvatar($userAvatar, $user);
         }
-
-        $this->imageRepository->uploadUserAvatar($userAvatar, $user);
 
         return $user;
     }
